Add Terraform node_group module and Puppetmaster project
Add initial Terraform code tom manage GOV.UK instances and
Puppetmaster project. We'll need to review the module to decide
how and where we want to manage ELB/ALB resources, but at least
we can start working on the Puppetmaster provisioning for now.

When using this code, for now the bootstrap ELB creted by the Puppetmaster
project needs to be added MANUALLY to the instance, because the
node_group module doesn't have a variable to add the launch_configuration
to additional ELBs.