Add EBS attach management to app-graphite
The app-graphite stack needs to be able to attach the data volume when
the instance is created the first time, or attach a previously formatted
volume if the instance is recreated. The first time a stack is created
the Puppet module takes care of creating and formating the LVM partition.

To manage additional EBS volumes we need an additional policy and user_data script.

The node_group module needs to be updated to output the name of the node
instance_iam_role, instead of the id, that is the value that we need if we
want to attach additional policies.