Update DNS infrastructure ADR
Update ADR to reflect the difference between a zone and a domain. We have
internal and external domains for each stack, external zones for each stack
(with a root zone delegating the stack subdomain to those zones), but a single root
internal zone, because we can't do delegation with internal domains.

Add another image to make the implementation easier to understand.