Update volume snippet and logs-elasticsearch
Update the attach-volume snippet to use a device tag so we can
attach several volumes to an instance and continue doing the
LVM management with Puppet.

The EBS TF resource needs to include a 'Device' tag so we can attach
the volume to the right device during the instance initialisation.

Puppet already manages the devices and LVM configuration, but we'll
need to update the lvm module in Puppet to use the puppetlabs version
instead of our fork, as it fixes several bugs that prevented a disk
already formatted from working properly when a new instance is running
Puppet the first time.