Add govuk-ssh wrapper
We will probably be working in multiple environments with multiple
stacks. I don't want to have to populate my SSH config with long lists
of the different stacks, so I wrote a wrapper which calls a script on
the jumpboxes to find the different machines I want to log into to.

This is an alternative to `ec2ssh`. It's main feature is that it does
not require maintaining a configuration file; just specify and go!