Add Mapit
Mapit is a slightly special case in that it's a cluster of two machines
that both run their own standalone PostgreSQL instances.

The uptime of these instances is important enough that they should have
their own volumes attached to them, although the amount of data is
relatively small on them.