Add app-router-backend
The router-backend instances are a slightly special case in that they
both host a MongoDB cluster and also an app to interact with the Mongo
cluster called "router-api".

This adds a cluster without dedicated EBS volumes on the assumption that
we take regular backups to maintain data.

To get around the MongoDB clustering issue, which entirely depends on
hostnames, we're using a separate ELB for each instance which we can
assign to Puppet.