Update app-api-postgres
The only database that is currently on api-postgresql database machines
is Stagecraft (performance platform). We should combine our PostgreSQL
instances so we only have to manage and maintain a single set.

Also update the name slightly to "postgresql" so it is more reflective
of the names we use across our Puppet code base (rather than the slang
"postgres").