Refactor Puppetmaster for Puppetserver
Refactor the Puppetmaster project to move to puppetserver and support
policy-based autosign with AWS certificate extensions:
- Add additional policy to create tags on instances when the client
certificate is signed
- Increase puppetmaster root volume size
- Update puppetmaster userdata snippet to remove unrestricted autosign, as
we are going to use policies. The policy script will be installed in govuk-puppet
- Move openjdk package installation to govuk-puppet and gpg related gems to
bootstrap script. govuk-puppet is going to ensure these gems are absent after
installing the puppetserver package, as puppetserver uses its own gem tool.