Set max_allowed_packet to 1GB for mysql in RDS
This creates a parameter group for the mysql-primary RDS instance that
sets the `max_allowed_packet` to 1GB. The default AWS value is 4MB
which means database restores of databases with large blobs (e.g.
whitehall) will fail (with `ERROR 2006 (HY000) at line 8500: MySQL
server has gone away`).

As the parameter group name needs to be added to the `db_instance`
resource this adds a new variable to the RDS module,
`parameter_group_name` which passes this value to both the instance and
replica resources..