Update frontend and calculators-frontend DNS names
This adds the required DNS records and certificates to allow the cache
instance to directly route requests to frontend and
calculators-frontend, essentially bypassing frontend-lb.

This is something we wanted to do later on, and it makes sense to do it
now because otherwise we will have conflicts in how we do DNS naming for
apps and services.

It's also worth noting I fixed a bug where the wrong protocol was used
with an LB listener, that meant that while requests were correctly being
routed to tcp/80, they were still encrypted so nginx was unable to
correctly parse the request.