Add variable to override default instance AMI
Currently the node_group module uses the `aws_ami` data source to determine the newest AMI ID
published by Ubuntu within a major version and use that in our launch configurations. This parameter
could be overriden in each project to set a specific image name, and this is going to be the needed
to avoid having to recreate the launch_configuration every time we get a new image patch.

This commit adds a new variable to all the projects to make sure we set the same image everywhere.
The default image version should be set in the common data file.