Add EBS volume to mirrorer
The mirrorer needs a large disk to store the static website prior to
upload. This adds a 100GB EBS volume to the instance.

This could, in theory, be done using a larger root volume but this would
mean that if the instance went away progress would be lost. This would
likely limit the usefulness of the static images if it had to repeat
large amounts of work.