Add external ELB to backend
Adding an external ELB to backend, along with all the service records
that originally existed on backend-lb, means that we can direct traffic
directly to the backend instances and remove the backend-lb tier.

Services speak to backend both internally and externally, so an internal
and external ELB are both required.