Use attached ENI rather than ELBs on mongo
We're seeing an issue where applications using the mongo cluster are
unable to connect to the cluster despite it being available.

We believe it may be related to the dynamic nature of ELBs, and the
mongo-ruby-driver currently being unable to handle DNS updates without a
reload of the application.

This commit removes the load balancers, creates an ENI and
attaches the ENI to the instance on boot using userdata.

As per the documentation, an IP associated with an ENI should follow the
interface to any attached instance:

http://docs.aws.amazon.com/AWSEC2/latest/UserGuide/using-eni.html

We can assign a DNS record to the IP address so the Puppet code
underneath does not have to change.

This requires adding additional userdata which finds the interface and
private IP using tags, and creates the configuration on the instance.

We have created a generic "reserved_ips" subnet incase we need to add
further ENIs to instances in the future.