Test Prometheus
Add some security rules and test instance to start installing
Prometheus. The internal ELB rules aren't properly set yet, this
configuration is just a placeholder. The test instance runs the `jumpbox`
profile (but not security rules) to test the installation and configuration
manually first.

We'll remove this project after the firebreak.