Add module alarms autoscaling
Add a new AWS module to manage autoscaling alarms. At the moment
it only includes an alarm on the number of instances in service,
but we could add more based on our needs. The alarm threshold can
be configured with a variable. The actions need to be specified
with an input variable.