Update RDS module for backups and replicas
Create some default values that will be assigned to an instance unless
specified otherwise:

 - skip_final_snapshot disabled by default (this should be enabled in
 Production. It doesn't hurt to include the name, so uncomment)
 - default backup_window of between 04:00 - 06:00
 - default maintenance_window of between 01:00 - 03:00 on a Monday
 - default backup_retention_period set to 7 days

The db_instance_replica has also been modified to remove the
allocated_storage parameter from the replica, as this is set by the
primary database instance.

In theory we should also be able to remove username, password, engine
and engine_version, but I got an error when I tried to apply this. All
these options are ignored when you specify a replicate_source_db.