Add AWS module alarms EC2
Add new alarms module to monitor EC2 instances.

It enables alarms for EC2 instances that belong to a specific
autoscaling group. For this reason it could be combined with the
autoscaling module, but it looks more tidy to organise modules per
metric namespace. Also, if we create a new intance_id parameter,
it could also serve to apply to individual instances instead of filtering
per ASG, which would make the module more generic.