Add AWS alarms ELB module
Add alarms module to monitor ELBs. We include 4 alarms based on HTTP
metrics that only make sense to enable on ELBs with HTTP listeners,
and other 2 alarms that can serve to monitor the status of ELBs with
TCP listeners. All the alarms can be disabled by setting the threshold
variable to 0.