Update RDS module for replicas
This commit makes a few changes to how the RDS module is configured.

Set some default options on the otherwise required paramaters so we
don't need to needlessly pass them in when creating a replica. It will
source these values from the primary it is replicating.

When creating a replica we also do not want to create another subnet
group, as it will use the primary instances subnet group.