Update outputs in module
Despite having the `join` workaround described, I was still not getting
any output related to the replica endpoint address, which meant that
creating the appropriate DNS record failed.

I've updated it to output the specific instances; if the instance is
created, then there will not be any outputs created with it. I think it
makes it clearer when addressing the outputs in the project itself.

I have tested this against projects without replicas without issue.