Use name_prefix instead of name for launch config
Terraform recommends in the
[documentation](https://www.terraform.io/docs/providers/aws/r/launch_configuration.html#using-with-autoscaling-groups)
to use launch configs in this way. We already set the
`create_before_destroy` lifecycle parameter, but because we hardcode the
name, terraform tries to create a new launch config with the same name
and fails. Updating the launch config to use name_prefix successfully
creates a new launch config with a different name and then applies that
to the ASG before destroying the existing launch config.

Because this is in a shared module, this will force a new launch config
when we next apply any changes using it, but this shouldnt result in any
difference except for the launch config name, which will have a random
suffix after it.