Change disks for Graphite
There have been common alerts for our graphite instance:

 - disk space
 - high disk time

We should increase the default disk space on the EBS volume, and also
remove setting the disk size on the default node_group as this is set
upstream by the module.

Also change the disk type to a disk capable of higher throughput:

See the [documentation](https://aws.amazon.com/ebs/details/) for
details.