Update Bouncer ELBs and rules
Bouncer should respond to traffic on port 80, as it handles redirects
that may not neccessarily be coming from an https request.

We should also ensure there is a certificate for https requests, and
update security groups so that the monitoring instance is able to test
the endpoint.

Add a security rule that explicitly allows traffic replay requests.