Allow monitoring to speak to Puppetdb
Someone had already added a DNS record for puppetdb, but we didn't allow
anything to access it.

We have a check which specifically uses puppetdb to check what nodes
require rebooting, and this is always run from the monitoring instance.

For now, only allow that instance access, and if we wish to open it up
in the future (to make it accessible with our fabric scripts, for
example), then we can do so.