Allow deploy instance to write to artefact bucket
Refactor the artefact bucket project to move policies to the policy
directory, and move the code into the main manifest.

I have not done this with the global replication and will get round to
this later.

This creates two policies: writer and reader.

These are separate polices incase we wish to allow reading from a
specific instance in the future.

Allow the deploy instance to write to the artefact bucket so it can make
a copy of artefacts into their version of the "deploy-to-environment"
branch. We use this branch to make assumptions about the version of the
app that is deployed, and will use it to automatically deploy apps when
nodes come up.

Router is pretty key to run on the cache instances, so we need to allow
this to happen.

Relates to https://github.com/alphagov/govuk-app-deployment/commit/636b1ac253a71d80cd6a6af1b02afe0adbe1fe3a