Add infra-monitoring outputs for individual SNS topics
We want to use SNS queues for several things: CloudWatch alarms,
ASG notifications and RDS notifications. We only had one SNS topic
to collect all these notifications, but in order to be able to
create subscriptions for individual notifications (for instance,
an email subscription for RDS event only), we need to sepatate
the SNS topics. We tried the filter subscription feature for a
shared SNS topic but this only works with message attributes
that can be added to custom messages, the previously mentioned
notifications don't have attributes.

For now we create three outputs to use with each type of notification
referencing the same SNS topic. We need to agree with GOV.UK how
they want to manage the notifications later, so at that point we
can decide if we want to have a single shared topic or several ones.