Use ELB ASG attachment
Use `aws_autoscaling_attachment` resources to attach an ASG to Classic
ELBs instead of the `load_balancers` parameter in the `aws_autoscaling_group`
resource.

The effect is similar in both cases, but both methods are not compatible and
we want to be able to create attachments to this ASG from other part of the
repository outside the `node_group` module. For more info:

https://www.terraform.io/docs/providers/aws/r/autoscaling_attachment.html