LB module configure health check path and add alarms
Our Vhosts default response for `/` is 444. We want to be able to
provide a good health check for our LB targets so we've been
updating Nginx to provide a 200 response code for `/_healthcheck`
requests.

This change sets the new default parameters for the health check and
also adds ALB alarms, similar to the ones we added in the alarms modules.