Update node_group implementation with instance_target_group_arns_length
Update the projects using the lb and node_group module to use
the new `instance_target_group_arns_length` variable.

To understand how this parameter gets calculated, we use
locals to define the association between listeners and target groups.

The number of target group is equal the unique values of the local
maps. If we use more than one map (for an external and internal LBs),
we need to add the those values.