Remove stackname from search domain
Hosts should not resolve to their own stack when doing searches because
we're providing resolution by the "top level" domain to assign which
services go where.

If we need something to go to a specific stack we can still specify the
FQDN, but a default lookup should resolve to the top level. For example,
if we want to connect to redis, we should be connecting by default to
`redis.integration.govuk-internal.digital`, rather than
`redis.blue.integration.govuk-internal.digital`.