Allow Jumpbox to PuppetDB
Our Fabric scripts are used by developers to run tasks. We should allow
them to have the full toolbox at their disposal as they migrate over.

If we wish to change how PuppetDB is accessed in the future we should
redesign it when we have time, rather than blocking other people do
work.

In terms of security, we are keeping with the same risk as the current
environment.