Revert "Revert "Remove stackname from search domain""
This reverts commit 0b787c121bfc42904eaf55b30cb20c03e48a7e6d.

I reverted because machines were unable to communicate with the
Puppetmaster. This is because infra-public-services should be deployed
both before and after instances are created.

Before to create a set of CNAMEs for the top level DNS, and after to
create the public load balancers.

I believe these items should be separated for "internal" and "external"
domains.