Refactor linter
I went down a rabbit hole in trying to add a way to add more rules.

This refactors the linter so we can add additional lints on the resource
names by adding rules in a YAML file.

Also includes tests for the method due to the logic. Ideally new types
of lint should be added to the rspec tests.