Add CloudTrail settings and alarms
At the moment Cloudtrail is configured manually to store log files in a
S3 bucket. With this change we add CloudTrail to the base infra-security
project to make sure it's managed properly. It has been configured to
store files in a S3 bucket and also send them to CloudWatch logs so we can
check quickly when needed and set some initial alarms.

In both cases (S3 and CloudWatch logs) we need to set the right roles and
permissions. For more information:

https://docs.aws.amazon.com/awscloudtrail/latest/userguide/logging-management-and-data-events-with-cloudtrail.html
https://docs.aws.amazon.com/awscloudtrail/latest/userguide/send-cloudtrail-events-to-cloudwatch-logs.html

We have configured 4 alarms to monitor the root account, unauthorised API calls
and logins, and authentication without MFA. The security team sent a recommendation
last year around CloudTrail that is also reflected in https://trello.com/c/ILcUMEts