Fix Bouncer LB health check
On the bouncer machines, Nginx is configured to use the 'bouncer'
vhost as the default one. The Puppet template does not allow to
insert custom code before the 'location /' block, so we can't set
our health check to '/_healthcheck' as we have done in the other
machines.

Bouncer already deploys a heath check on `/healthcheck`, so use
that instead.