Add script for Jenkins deploy
This script invokes the `build-terraform-project.sh` script and should
be triggered using Jenkins.

Sops is a precursor as an install to run these scripts, and should be
managed by Puppet as it is less liable to change than Terraform.

If the Terraform version is not specified using TERRAFORM_VERSION, then
it will use the default installed on the system where it is being run.

If you wish to test newer versions of Terraform (quite a common
occurance at this point in the Terraform lifecycle), then you can
specify a version and the script will download, verify and set the path
for that version of Terraform.