Add new govukcli wrapper tool
This is to help developers SSH to the environment, and also remove the
older govuk-ssh tool which has a proxy script which can also be removed
from Puppet.

I've written it so it can be expanded for other uses in the future.

Originally we wanted to write something in Ruby as a gem and have
interactions with the AWS API, but didn't get a chance to do it.

I wrote in the concept of "contexts" (nicking kubectl language) so
people don't have to fiddle around and maintain SSH config.

Set a context with "set-context", find out the current context with
"get-context" and view available contexts with "list-contexts".

With SSH to further enable no SSH config, we can set a specific username
to use:

ssh set-user lauramartin

Otherwise it will use the default shell user.