infra-security support empty user list
As we add more environments and user roles, sometimes we don't
need to create all of them on each environment. For instance,
the users group doesn't need to be present on the Test environment,
where we want people to be able to create more things.

To support empty lists of users from the data files, the `role_X_user_arns`
parameters in the infra-security manifest need to default to empty list,
and the `role_user` module needs to only create a role when the list
of users is not null. This is needed because the `Principal` section in
the policy can't be an empty string.