Allow backwards compatability in govukcli tool
This allows connecting to the current environments in Carrenza:

 - CI
 - Staging
 - Production

It also adds logic to add to the new AWS environments:

 - Staging
 - Production

When connecting to the environments in Carrenza, it adds ability to SSH
to specific machines like we're used to doing. So you can do stuff like:

```
govukcli set-context ci
govukcli ssh licensing-frontend-1
```

This helps people get used to working in a single way with a single
tool, and without having to manage a bunch of SSH config.