Allow option to list node-types
Someone suggested a way to list node types. In Carrenza, this is tricky,
but since I built in the option to SSH straight to specific machines
it's less valuable.

In AWS, we can use the tags assigned to instances, so using a bit of jq
magic, we can list all the types of instances available to SSH to.