Allow backend machines access to graphite
Testing out some changes to the release app on integration I
found that it timed out when trying to access a grafana 
dashboard, but on staging and production it did not.  It times
out in the "open" phase, rather than the "read" phase of the
http request, so we figured it was a firewall issue.  Opening
up access to graphite machines to the backend machines
(which is where the release app lives) should mean that the
request won't time out any more.