Promote Deploy-App Content Between Environments
Necessity
---------
We have an existing process where a Jenkins job called **Deploy_Apps**
creates a Go language binary file under an S3 bucket for the relevant
environment. For example **govuk-integration-artefact**. When we build
another environment, for example **Staging**, then that binary file
should be available in the environment specific bucket for the new
build. For staging , it will be  **govuk-staging-artefact**. Hence, we
need a mechanism to promote the binary file accross environments.

Concept
-------
We create a mechanism that will be aware when a new file gets created.
1. We will create an "SNS topic" in the environment where the binary
file will be created.
2. We create an "S3 bucket notification event and associate it to the
bucket that will get the new file.
3. The S3 bucket event will be associated to the "SNS topic".

We create a mechanism that can know when the new file is created and
retrieve the file.
1. We create a "Lambda" function that can copy a file from another "S3"
bucket.
2. We create an "SNS Subscription" that subscribes to the "SNS Topic" at
the source.
3. We associate the "SNS Subscription" to the "Lambda" function.

When a new file gets created, the following will happen.(We will use
"Integration"(INT) as source and "Staging"(STG) as destination)
New file -> INT SNS Topic notified -> STG Subscription alerted -> STG
Lambda function syncs the file

Working Components
------------------

Lambda Files
------------
The lambda related files are stored in terraform/lambda/ArtefactSync.

Terraform Resources
-------------------
There are number of terraform resources that work together to complete
the functionality.

govuk-artefact-bucket-policy
----------------------------
This resource assigns an S3 bucket policy. The policy will restrict
access and allow the AWS Integration and Staging account access to the
artefact bucket.

artefact_topic
--------------
This resource creates an AWS SNS Topic in the source environment.

artefact_topic_policy and artefact_sns_topic_policy
---------------------------------------------------
artefact_sns_topic_policy  is the actual policy that give Lambda related
permissions to interact with AWS SNS. artefact_topic_policy acts like a
policy attachment to the SNS topic.

artefact_bucket_notification
----------------------------
This resource links the "S3 Bucket Event" to the "SNS Topic".

artefact_topic_subscription
---------------------------
This resource is the "SNS Subscription" at the target location. Also,
this resource will associate to the "Lambda Function" at the source
location.

artefact_lambda_function
-------------------------
This resource creates the Lambda function and connects to the relevent
IAM role.

govuk_artefact_lambda_role
--------------------------
This resource allows the Lambda function to assume the relevant IAM
role.

govuk_artefact_policy
---------------------
This resource will create an IAM policy. This policy will allow logs to
be written to CloudWatch, during the Lambda function's activity.

govuk_artefact_policy_attachment
--------------------------------
This resource is used to attach the "IAM Role" and "IAM Policy" used by
the Lambda function.

govuk_artefact_lambda_sns
-------------------------
This is the resource that allows the "SNS Subscription" to trigger the
"Lambda Function" and cause the execution.

Solo: @suthagarht