Add verbose error for no backend file
When a backend file is missing, a misleading error is shown. We should
produce an error that guides the user to try the correct thing. This
should produce a list of possible stacks to deploy to if the backend
file does not exist for that stack in the specified environment.