Add completion for govukcli
Someone mentioned about bash completion, so as a firebreak task I
thought I'd add it to govukcli as it makes it easier to get to
instances.

The `$lastword` variable is useful as it stops recursive completion
against commands, which is why it's use in all the commands across these
gross case statements. Writing in case statements does make it easier to
add in further options in the future, though.