AWS S3 Backup cross region replication
- We are aattempting to replicate S3 database backup bucket accross to
  another region.

- Reference -
  https://www.terraform.io/docs/providers/aws/r/s3_bucket.html

New Resources
------------

- database_backups_replica
--------------------------
This resource was used to create the replica bucket in the secondary
region.

- s3_backup_replica_assume_role_template & backup_replication_role
------------------------------------------------------------------
The template resource points to the template file that contains the
assume role policy used by S3 replication service. The role resource
creates the AWS IAM role.

- s3_backup_replica_policy_template & s3_backup_replica_policy
--------------------------------------------------------------
The template resource points to the policy file that defines specific
permissions that permits the replication service to access a specific S3
bucket and it's contents. The policy resource creates the actual policy.

- backup_replication_policy_attachment
--------------------------------------
This resource will attach the policy to the role.

Pair: @suthagarht @schmie