Manage Fastly service for data.gov.uk in Terraform
This is the basic definition for data.gov.uk's Fastly service. The current resource has been imported, and the definition written so that `plan` results in no changes. Soon, I'll be changing this so it has different backends for different paths. There's only a `production` backend as there's only a Production service.