AWS S3 Database Backup Writer
- We had an AWS IAM policy called "database_backups_writer". This policy
provided write access to the "govuk-production-database-backups" bucket.
This policy was attached to the instances that needed access to write/read content from the bucket. There were no restrictions in place to limit which objects can be accessed and which environment has read/write access..

- The requirement has changed. We want the AWS Production environment instances to be able to write to the object called "govuk-production-database-backups/destination".
We also need to provide read access to this location for the AWS  staging and integration environments.

- To accomplish this, we have modified the "writer" policy. The modified
policy will define a single resource per project. This enables us to
limit which objects within the bucket can be accessed (prefix). We are
using the prefix value in the backup script. https://github.com/alphagov/govuk-env-sync/blob/mongo-to-s3-bash/mongo-to-s3.sh#L45

- We have also introduced a condition inside the terraform projects.
This condition defines that the writer policy will only be attached to AWS
production environment instances. The other environments will get a
reader policy attached.

https://trello.com/c/H3CJpNje/1257-data-sync-aws-s3-backup-bucket-access

Solo: @suthagarht