Glue crawler and table for assets CDN Logs
This sets up a table which correlates with the logging pattern applied
to fastly which is:

%h\t%u\t%{%Y-%m-%d %H:%M:%S.}t%{msec_frac}t\t%{%z}t\t%m\t%{req.url}V\t%>s\t%{time.elapsed.sec}V.%{time.elapsed.msec_frac}V\t%{time.to_first_byte}V\t%B\t%{Content-Type}o\t%{User-Agent}i\t%{Fastly-Backend-Name}o\t%{server.datacenter}V\t%{if(resp.http.X-Cache ~"HIT", "HIT", "MISS")}V\t%{tls.client.protocol}V\t%{tls.client.cipher}V

The components of this is explained within the code.

I had initially wanted to use JSON for this as this would make it far
easier to add/remove columns in the future but I hit a deadly bug that
any strange characters would break through fastly `cstr_escape` function
and leave me with broken JSON.

The crawler is configured to run every 4 hours and to maintain the
schema that is defined here rather than any changes in the data.