New VPC resources for VPN connection to Carrenza
These new resources should setup the gateways and VPN connection for creating a link between our VPC and the Carrenza environment. There are a couple of new variables that need to be defined for each environment. We should be able to get those from Carrenza. AWS then generates a VPN configuration which we can get from the console and use to configure the VCloud networking in 'govuk-provisioning' or however we want to do that bit.