New VPC resources for VPN to Carrenza.
This creates the two gateways and a VPN for connecting the VPC to Carrenza. There are 2 new variables to add in govuk-aws-data also. This should create the VPN and then we can get the created configuration to use to update the VCloud configuration in govuk-provisioning to deploy the Carrenza side of things.