Do not group Athena query results by request_received
This causes problems around midnight, where a request can be received
on one day but be logged for the next, if the duration of the request
pushed it over the day change.  The SELECT hard-codes the date, rather
than returning the request_received, so we see two apparently
identical requests on the same day.

We don't really need the data to be as precise as possible, so we'll
just say that all requests which finished processing on a given day
were received on that day.  The error introduced in Day N by pushing
some requests on Day N+1 is resolved by some requests on Day N+1
inevitably being pushed to Day N+2, and so on.