GovUK Migr Trello 1646: Deny EIP Release
Adding Policy to make sure they are unable to release/remove
EIPs from the environments.

The reason we wish to deny removal of the EIPs, is that we do not want
the EIPs changing. These EIPs are attahced to teh NAT gateways and if we
did have to re-spin the nat gateways then we would re-attach these EIPs
again.

This needs to be merged with branch https://github.com/alphagov/govuk-aws-data/tree/GovUK-Migr-Trello-1646

Pair: @ronocg @th31nitiate