BAU: Fixup Terraform resources for VPN
The VPN resources / security group will not exist in integration_aws but
will exist in staging.

This PR is to put in place a count loop for the vpn security group and
rules, so that terraform will not attempt to create them for
environments that do not have the correct variable set.

Solo: @ronocg <conor.glynn@digital.cabinet-office.gov.uk>