Trello-1808: Tie down access to content-stores AWS
The content-store external ALB and external ELB in staging_aws and
production_aws are currently opened to the world. Here we are limiting
access to the content-stores from the corresponding environment in
Carrenza. This is so teh frontends in Carrenza which have not been
migrated yet are allowed to talk with the content-store.

We are allowing the GDS office ip ranges in also.

Solo: @ronocg <conor.glynn@digital.cabinet-office.gov.uk>